fix(sale): compose full ticket section from subZone + filter /tickets by section
The Balcony-side fix in TicketReceivedV2 never worked: the public API
overwrites `position` with `seat_number` when present, so the client got
"39" instead of "balconyright". The raw position travels in `subZone`;
new shared helper src/lib/ticketSection.ts uses it, applied to both
TicketPublicV2 (the mail-linked view, which had no fix at all) and
TicketReceivedV2.

Also stop fetching all ~1000 event tickets on every section open in the
seat picker: all seats in a group share `position`, and the backend
already supports `?position=` on /public/events/{id}/tickets, so
getPriceTiers now filters to the opened section.
